refactor: move github function
