+++ src/commands/basemaps-github/make.cog.github.ts
@@ -10,7 +10,7 @@ import { fsa } from '@chunkd/fs';
 import { logger } from '../../log.js';
 import { DEFAULT_PRETTIER_FORMAT } from '../../utils/config.js';
 import { prettyPrint } from '../format/pretty.print.js';
-import { createPR, GithubApi } from './github.js';
+import { createPR, GithubApi } from '../../utils/github.js';
 
 export enum Category {
   Urban = 'Urban Aerial Photos',
